+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 1.0.0+1
+
+- Added `enabled` parameter to `l10n_decompose`.
+```yaml
+l10n_decompose:
+  # Optional; If not specified, defaults to true.
+  enabled: true
+```
+
 ## 1.0.0
 
 - Initial version.